+++ core/src/daniele/comunication/LocalActor.scala
@@ -1,16 +1,20 @@
 package daniele.comunication
 
 import akka.actor.{Props, UntypedAbstractActor}
-import daniele.utils.IntMsg
+import daniele.comunication.Messages.{IntMsg, StringMsg}
+
 
 
 class LocalActor(name: String) extends UntypedAbstractActor {
 
   override def onReceive(message: Any): Unit = message match {
-    case x: IntMsg =>
-      println(name + ": " + x.getVal + " from: " + sender())
-      x inc()
-      sender().tell(new IntMsg(x.getVal), getSelf())
+    case msg: IntMsg =>
+      println(name + ": " + msg.v + " from: " + sender())
+      //x inc()
+      sender().tell(IntMsg(msg.v+1), getSelf())
+    case msg: StringMsg =>
+      println("message: " + msg.v)
+      sender().tell(StringMsg("received"), getSelf())
     case _ => println("message unknown")
   }
 }
