+++ core/src/daniele/comunication/Messages.scala
@@ -1,0 +1,6 @@
+package daniele.comunication
+
+object Messages {
+  case class IntMsg(v: Int)
+  case class StringMsg(v: String)
+}
